Finished

+++ Currency/src/main/java/org/it_academy/currency/dao/CurrencyDao.java
@@ -1,9 +1,10 @@
 package org.it_academy.currency.dao;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.PersistenceException;
 import org.it_academy.currency.api.CRUD.ICRUDHibernate;
 import org.it_academy.currency.dao.entity.Currency;
+import org.it_academy.currency.exceptions.CurrencyServiceException;
 
 import java.util.List;
 
@@ -18,7 +19,7 @@ public class CurrencyDao implements ICRUDHibernate {
 
     @Override
     public long save(Currency currency) {
-        EntityManager entityManager = manager.getEntityManager();
+        EntityManager entityManager = manager.getManager();
 
         entityManager.getTransaction().begin();
 
@@ -32,13 +33,15 @@ public class CurrencyDao implements ICRUDHibernate {
 
     @Override
     public void delete(Currency currency) {
-        EntityManager entityManager = manager.getEntityManager();
+        EntityManager entityManager = manager.getManager();
 
         entityManager.getTransaction().begin();
 
         Currency currency1 = entityManager.find(Currency.class, currency.getId());
 
-        entityManager.remove(currency1);
+        if (currency1 != null) {
+            entityManager.remove(currency1);
+        }
 
         entityManager.getTransaction().commit();
         entityManager.close();
@@ -46,12 +49,13 @@ public class CurrencyDao implements ICRUDHibernate {
 
     @Override
     public List<Currency> getAll() {
-        EntityManager entityManager = manager.getEntityManager();
+        EntityManager entityManager = manager.getManager();
 
         entityManager.getTransaction().begin();
 
-        List<Currency> currencies = (List<Currency>) entityManager.createQuery("SELECT c FROM Currency c")
-                .getResultList();
+        List<Currency> currencies = (List<Currency>) entityManager.createQuery(
+                "SELECT c FROM Currency c ORDER BY id"
+                ).getResultList();
 
         entityManager.getTransaction().commit();
         entityManager.close();
@@ -61,7 +65,7 @@ public class CurrencyDao implements ICRUDHibernate {
 
     @Override
     public Currency get(long id) {
-        EntityManager entityManager = manager.getEntityManager();
+        EntityManager entityManager = manager.getManager();
 
         entityManager.getTransaction().begin();
 
@@ -78,11 +82,17 @@ public class CurrencyDao implements ICRUDHibernate {
 
     @Override
     public void update(Currency currency) {
-        EntityManager entityManager = manager.getEntityManager();
+        EntityManager entityManager = manager.getManager();
 
         entityManager.getTransaction().begin();
+        Currency currency1 = entityManager.find(Currency.class, currency.getId());
 
-        entityManager.merge(currency);
+        if (currency1 == null) {
+            throw new CurrencyServiceException(404, "Not found");
+        }
+
+        currency1.setDescription(currency.getDescription());
+        currency1.setUpdateDate(currency.getUpdateDate());
 
         entityManager.getTransaction().commit();
         entityManager.close();
